+++ src/AdvancedExecutableInterface.ts
@@ -11,7 +11,7 @@ export type AdvancedExecutableConfig = {
   name?: ResolvableString;
   description?: ResolvableString;
   comment?: string;
-  tags?: string[];
+  testTags?: string[];
   cwd?: ResolvableString;
   env?: Record<string, ResolvableString>;
   envFile?: ResolvableString;