+++ src/ConfigOfExecGroup.ts
@@ -36,7 +36,7 @@ export class ConfigOfExecGroup implements vscode.Disposable {
     private readonly _exclude: string | undefined,
     private readonly _name: string | undefined,
     private readonly _description: string | undefined,
-    _tags: string[], // TODO: pass down to tests and set on them if speficied
+    private readonly _testTags: string[],
     private readonly _cwd: string,
     private readonly _env: { [prop: string]: string },
     private readonly _envFile: string | undefined,
@@ -458,6 +458,7 @@ export class ConfigOfExecGroup implements vscode.Disposable {
       this._shared,
       this._name,
       this._description,
+      this._testTags.map(t => new vscode.TestTag(t)),
       filePath,
       {
         cwd: resolvedCwd,